+++ multistock-backend/app/Http/Controllers/MercadoLibre/Credentials/getAllCredentialsDataController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers\MercadoLibre\Reportes;
+namespace App\Http\Controllers\MercadoLibre\Credentials;
 
 use App\Models\MercadoLibreCredential;
 use Illuminate\Support\Facades\Http;
